+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Add `llms_txt_max_lines` configuration option to limit `llms-full.txt` file size
 - Automatically add content from `.. include::` directives
+- Add path resolution for certain directives
 
 0.1.0
 -----